Fix type errors and refactor client components
- Fix FeedProduct nutritional_specs type alignment
- Fix AllAgroChemicalsClient useQuery initialData type
- Extract client components for interactive features
