+++ apps/admin-fnp/lib/schemas.ts
@@ -934,7 +934,7 @@ export const FeedProductSchema = z.object({
     id: z.string(),
     feed_product_id: z.string(),
     nutritional_spec_id: z.string(),
-    name: z.string().optional(),
+    name: z.string().default(""),
     value: z.string(),
     unit: z.string(),
     qualifier: z.string(),
