+++ apps/client-fnp/app/(landing)/agrochemical-guides/[category]/page.tsx
@@ -1,14 +1,6 @@
-"use client"
-
-import { use } from "react"
-import { useQuery } from "@tanstack/react-query"
 import { queryAgroChemicalsByCategory } from "@/lib/query"
-import { Button } from "@/components/ui/button"
-import { Beaker } from "lucide-react"
-import { AgroChemicalFilterSidebar } from "@/components/generic/agroChemicalFilterSidebar"
-import { AgroChemicalCard } from "@/components/agrochemical/AgroChemicalCard"
-import { useQueryStates, parseAsArrayOf, parseAsString, parseAsInteger } from "nuqs"
 import Link from "next/link"
+import { AgroCategoryClient } from "./AgroCategoryClient"
 
 interface CategoryPageProps {
     params: Promise<{
@@ -16,35 +8,30 @@ interface CategoryPageProps {
     }>
 }
 
-export default function AgroChemicalCategoryPage({ params }: CategoryPageProps) {
-    const { category } = use(params)
+export default async function AgroChemicalCategoryPage({ params }: CategoryPageProps) {
+    const { category } = await params
 
-    const [queryState, setQueryState] = useQueryStates({
-        brand: parseAsArrayOf(parseAsString),
-        target: parseAsArrayOf(parseAsString),
-        active_ingredient: parseAsArrayOf(parseAsString),
-        p: parseAsInteger.withDefault(1),
-    })
+    let initialChemicals: any[] = []
+    let initialTotal = 0
 
-    const { data: chemicalsData, isLoading: chemicalsLoading } = useQuery({
-        queryKey: ["agrochemicals-category", category, queryState.p, queryState.brand, queryState.target, queryState.active_ingredient],
-        queryFn: () => queryAgroChemicalsByCategory({
+    try {
+        const response = await queryAgroChemicalsByCategory({
             category,
-            p: queryState.p,
-            brand: queryState.brand || [],
-            target: queryState.target || [],
-            active_ingredient: queryState.active_ingredient || [],
-        }),
-        refetchOnWindowFocus: false,
-    })
-
-    const chemicals = chemicalsData?.data?.data || []
-    const totalPages = Math.ceil((chemicalsData?.data?.total || 0) / 20)
-    const categoryName = chemicals[0]?.agrochemical_category?.name || category
-
-    const handlePageChange = (newPage: number) => {
-        setQueryState({ p: newPage })
+            p: 1,
+            brand: [],
+            target: [],
+            active_ingredient: [],
+        })
+        initialChemicals = response?.data?.data || []
+        initialTotal = response?.data?.total || 0
+    } catch (error) {
+        console.error("Error fetching agrochemicals by category:", error)
     }
+
+    const categoryName = initialChemicals[0]?.agrochemical_category?.name || category
+        .split('-')
+        .map((word: string) => word.charAt(0).toUpperCase() + word.slice(1))
+        .join(' ')
 
     return (
         <div className="min-h-screen bg-gradient-to-b from-background to-muted/20">
@@ -72,107 +59,12 @@ export default function AgroChemicalCategoryPage({ params }: CategoryPageProps) 
                     </p>
                 </div>
 
-                <div className="flex flex-col lg:flex-row gap-8">
-                    {/* Sidebar Filters */}
-                    <aside className="w-full lg:w-64 flex-shrink-0">
-                        <AgroChemicalFilterSidebar hideCategory={true} />
-                    </aside>
-
-                    {/* Main Content */}
-                    <main className="flex-1">
-                        {chemicalsLoading ? (
-                            <div className="grid grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-5">
-                                {[...Array(8)].map((_, i) => (
-                                    <div key={i} className="animate-pulse">
-                                        <div className="bg-card border border-border rounded-lg overflow-hidden">
-                                            <div className="aspect-square bg-muted" />
-                                            <div className="p-4 space-y-3 border-t">
-                                                <div className="h-3 bg-muted rounded w-1/3" />
-                                                <div className="h-4 bg-muted rounded w-4/5" />
-                                                <div className="h-4 bg-muted rounded w-3/5" />
-                                                <div className="flex gap-4 pt-2 border-t">
-                                                    <div className="h-3 bg-muted rounded w-16" />
-                                                    <div className="h-3 bg-muted rounded w-16" />
-                                                </div>
-                                            </div>
-                                        </div>
-                                    </div>
-                                ))}
-                            </div>
-                        ) : chemicals.length === 0 ? (
-                            <div className="text-center py-12">
-                                <Beaker className="w-16 h-16 mx-auto text-muted-foreground mb-4" />
-                                <p className="text-muted-foreground mb-4">No {categoryName.toLowerCase()} found matching your filters.</p>
-                                <Link href="/agrochemical-guides/all">
-                                    <Button variant="outline">View All Agrochemicals</Button>
-                                </Link>
-                            </div>
-                        ) : (
-                            <>
-                                <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-5">
-                                    {chemicals.map((chemical: any) => (
-                                        <AgroChemicalCard
-                                            key={chemical.id}
-                                            chemical={chemical}
-                                            mode="guide"
-                                        />
-                                    ))}
-                                </div>
-
-                                {/* Pagination */}
-                                {totalPages > 1 && (
-                                    <div className="mt-8 flex justify-center gap-1">
-                                        <Button
-                                            variant="outline"
-                                            size="sm"
-                                            onClick={() => handlePageChange(Math.max(1, queryState.p - 1))}
-                                            disabled={queryState.p === 1}
-                                        >
-                                            Previous
-                                        </Button>
-                                        {Array.from({ length: totalPages }, (_, i) => i + 1)
-                                            .filter(pageNum => {
-                                                // Show first page, last page, current page, and pages around current
-                                                return (
-                                                    pageNum === 1 ||
-                                                    pageNum === totalPages ||
-                                                    (pageNum >= queryState.p - 2 && pageNum <= queryState.p + 2)
-                                                )
-                                            })
-                                            .map((pageNum, idx, arr) => {
-                                                const prevPageNum = arr[idx - 1]
-                                                const showEllipsis = prevPageNum && pageNum - prevPageNum > 1
-
-                                                return (
-                                                    <div key={pageNum} className="flex items-center gap-1">
-                                                        {showEllipsis && (
-                                                            <span className="px-2 text-muted-foreground">...</span>
-                                                        )}
-                                                        <Button
-                                                            variant={queryState.p === pageNum ? "default" : "outline"}
-                                                            size="sm"
-                                                            onClick={() => handlePageChange(pageNum)}
-                                                            className="min-w-[40px]"
-                                                        >
-                                                            {pageNum}
-                                                        </Button>
-                                                    </div>
-                                                )
-                                            })}
-                                        <Button
-                                            variant="outline"
-                                            size="sm"
-                                            onClick={() => handlePageChange(queryState.p + 1)}
-                                            disabled={queryState.p >= totalPages}
-                                        >
-                                            Next
-                                        </Button>
-                                    </div>
-                                )}
-                            </>
-                        )}
-                    </main>
-                </div>
+                <AgroCategoryClient
+                    category={category}
+                    categoryName={categoryName}
+                    initialChemicals={initialChemicals}
+                    initialTotal={initialTotal}
+                />
             </div>
         </div>
     )
