+++ apps/client-fnp/app/(landing)/animal-health-guides/all/page.tsx
@@ -1,150 +1,28 @@
-"use client"
+import { BaseURL } from "@/lib/schemas"
+import { AllAnimalHealthClient } from "./AllAnimalHealthClient"
 
-import { useQuery } from "@tanstack/react-query"
-import { queryAllAnimalHealthProducts } from "@/lib/query"
-import { Button } from "@/components/ui/button"
-import { Beaker } from "lucide-react"
-import { AnimalHealthFilterSidebar } from "@/components/generic/animalHealthFilterSidebar"
-import { AnimalHealthCard } from "@/components/animalhealth/AnimalHealthCard"
-import { useQueryStates, parseAsArrayOf, parseAsString, parseAsInteger } from "nuqs"
+const fetchOptions: RequestInit = process.env.NODE_ENV === "production"
+    ? { next: { revalidate: 3600 } } as RequestInit
+    : { cache: "no-store" }
 
-export default function AllAnimalHealthPage() {
-    const [queryState, setQueryState] = useQueryStates({
-        brand: parseAsArrayOf(parseAsString),
-        target: parseAsArrayOf(parseAsString),
-        active_ingredient: parseAsArrayOf(parseAsString),
-        used_on: parseAsArrayOf(parseAsString),
-        p: parseAsInteger.withDefault(1),
-    })
-
-    const { data: productsData, isLoading: productsLoading } = useQuery({
-        queryKey: ["animal-health-all", queryState.p, queryState.brand, queryState.target, queryState.active_ingredient, queryState.used_on],
-        queryFn: () => queryAllAnimalHealthProducts({
-            p: queryState.p,
-            brand: queryState.brand || [],
-            target: queryState.target || [],
-            active_ingredient: queryState.active_ingredient || [],
-            used_on: queryState.used_on || [],
-        }),
-        refetchOnWindowFocus: false,
-    })
-
-    const products = productsData?.data?.data || []
-    const totalPages = Math.ceil((productsData?.data?.total || 0) / 20)
-
-    const handlePageChange = (newPage: number) => {
-        setQueryState({ p: newPage })
+async function getAllProducts() {
+    try {
+        const res = await fetch(`${BaseURL}/animalhealth/all`, fetchOptions)
+        if (!res.ok) return { data: [], total: 0 }
+        const json = await res.json()
+        return { data: json?.data || [], total: json?.total || 0 }
+    } catch {
+        return { data: [], total: 0 }
     }
+}
+
+export default async function AllAnimalHealthPage() {
+    const { data: products, total } = await getAllProducts()
 
     return (
         <div className="min-h-screen bg-gradient-to-b from-background to-muted/20">
             <div className="mx-auto max-w-7xl px-6 lg:px-8 py-12">
-                {/* Header */}
-                <div className="mb-8">
-                    <h1 className="text-4xl font-bold tracking-tight font-heading mb-4">
-                        All Animal Health Products
-                    </h1>
-                    <p className="text-lg text-muted-foreground">
-                        Browse our complete collection of animal health products for poultry and livestock
-                    </p>
-                </div>
-
-                <div className="flex flex-col lg:flex-row gap-8">
-                    {/* Sidebar Filters */}
-                    <aside className="w-full lg:w-64 flex-shrink-0">
-                        <AnimalHealthFilterSidebar />
-                    </aside>
-
-                    {/* Main Content */}
-                    <main className="flex-1">
-                        {productsLoading ? (
-                            <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-5">
-                                {[...Array(6)].map((_, i) => (
-                                    <div key={i} className="animate-pulse">
-                                        <div className="bg-card border border-border rounded-lg overflow-hidden">
-                                            <div className="aspect-square bg-muted" />
-                                            <div className="p-4 space-y-3 border-t">
-                                                <div className="h-3 bg-muted rounded w-1/3" />
-                                                <div className="h-4 bg-muted rounded w-4/5" />
-                                                <div className="h-4 bg-muted rounded w-3/5" />
-                                                <div className="flex gap-4 pt-2 border-t">
-                                                    <div className="h-3 bg-muted rounded w-16" />
-                                                    <div className="h-3 bg-muted rounded w-16" />
-                                                </div>
-                                            </div>
-                                        </div>
-                                    </div>
-                                ))}
-                            </div>
-                        ) : products.length === 0 ? (
-                            <div className="text-center py-12">
-                                <Beaker className="w-16 h-16 mx-auto text-muted-foreground mb-4" />
-                                <p className="text-muted-foreground">No animal health products found matching your filters.</p>
-                            </div>
-                        ) : (
-                            <>
-                                <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-5">
-                                    {products.map((product: any) => (
-                                        <AnimalHealthCard
-                                            key={product.id}
-                                            product={product}
-                                        />
-                                    ))}
-                                </div>
-
-                                {/* Pagination */}
-                                {totalPages > 1 && (
-                                    <div className="mt-8 flex justify-center gap-1">
-                                        <Button
-                                            variant="outline"
-                                            size="sm"
-                                            onClick={() => handlePageChange(Math.max(1, queryState.p - 1))}
-                                            disabled={queryState.p === 1}
-                                        >
-                                            Previous
-                                        </Button>
-                                        {Array.from({ length: totalPages }, (_, i) => i + 1)
-                                            .filter(pageNum => {
-                                                return (
-                                                    pageNum === 1 ||
-                                                    pageNum === totalPages ||
-                                                    (pageNum >= queryState.p - 2 && pageNum <= queryState.p + 2)
-                                                )
-                                            })
-                                            .map((pageNum, idx, arr) => {
-                                                const prevPageNum = arr[idx - 1]
-                                                const showEllipsis = prevPageNum && pageNum - prevPageNum > 1
-
-                                                return (
-                                                    <div key={pageNum} className="flex items-center gap-1">
-                                                        {showEllipsis && (
-                                                            <span className="px-2 text-muted-foreground">...</span>
-                                                        )}
-                                                        <Button
-                                                            variant={queryState.p === pageNum ? "default" : "outline"}
-                                                            size="sm"
-                                                            onClick={() => handlePageChange(pageNum)}
-                                                            className="min-w-[40px]"
-                                                        >
-                                                            {pageNum}
-                                                        </Button>
-                                                    </div>
-                                                )
-                                            })}
-                                        <Button
-                                            variant="outline"
-                                            size="sm"
-                                            onClick={() => handlePageChange(queryState.p + 1)}
-                                            disabled={queryState.p >= totalPages}
-                                        >
-                                            Next
-                                        </Button>
-                                    </div>
-                                )}
-                            </>
-                        )}
-                    </main>
-                </div>
+                <AllAnimalHealthClient initialProducts={products} initialTotal={total} />
             </div>
         </div>
     )
